[EGERIA 22] Asset Catalog OMAS - API module added + refactoring
