+++ open-metadata-implementation/access-services/asset-catalog/README.md
@@ -16,3 +16,18 @@ to it.  This includes:
 * glossary terms
 * schema elements
 * classifications
+
+The Asset Catalog REST API supports:
+
+* the retrieval of assets based on unique identifiers
+* the retrieval of asset's relationships and classifications
+* the retrieval of assets based on known classification or relationship
+* to query for related assets and to retrieve an asset neighborhood
+
+The module structure for the Asset Catalog OMAS is as follows:
+
+* [asset-catalog-client](asset-catalog-client) supports the client library.
+* [asset-catalog-api](asset-catalog-api) supports the common Java classes that are used both by the client and the server.
+* [asset-catalog-server](asset-catalog-server) supports in implementation of the access service and its related event management.
+* [asset-catalog-spring](asset-catalog-spring) supports the REST API using the [Spring]__(../../../developer-resources/Spring.md)__ libraries.
+
