+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-api/README.md
@@ -1,0 +1,8 @@
+<!-- SPDX-License-Identifier: Apache-2.0 -->
+
+# Asset Catalog Open Metadata Access Service (OMAS) API
+
+The Asset Catalog OMAS supports a REST API for requests for asset metadata.
+
+The Asset Catalog OMAS API provides the shared Java classes between the
+server and client.
