+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-api/pom.xml
@@ -14,26 +14,23 @@
 
     <modelVersion>4.0.0</modelVersion>
 
-    <name>Asset Catalog OMAS Server-side</name>
+    <name>Asset Catalog OMAS API</name>
     <description>
-        Server-side support for the Asset Catalog Open Metadata Access Service (OMAS).
+        API classes for the Asset Catalog Open Metadata Access Service (OMAS).
     </description>
 
-    <artifactId>asset-catalog-server</artifactId>
+    <artifactId>asset-catalog-api</artifactId>
 
+    <properties>
+        <lombok.version>1.16.20</lombok.version>
+    </properties>
+    
     <dependencies>
         <dependency>
-            <groupId>org.odpi.egeria</groupId>
-            <artifactId>asset-catalog-api</artifactId>
-            <version>${open-metadata.version}</version>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>${lombok.version}</version>
+            <scope>provided</scope>
         </dependency>
-
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-web</artifactId>
-            <version>${spring-web.version}</version>
-        </dependency>
-
     </dependencies>
-
 </project>
