+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-api/src/main/java/org/odpi/openmetadata/accessservice/assetcatalog/model/Endpoint.java
@@ -6,32 +6,23 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 
-import java.util.Date;
-
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.NONE;
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.PUBLIC_ONLY;
 
 /**
- * Asset object holds properties that are used for displaying details of an asset.
+ * The Endpoint object contains the properties needed to create a connection in order to access a specific data assets.
  */
 @JsonAutoDetect(getterVisibility = PUBLIC_ONLY, setterVisibility = PUBLIC_ONLY, fieldVisibility = NONE)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Data
-public class Asset {
+public class Endpoint {
 
-    private String GUID;
-    private String metadataCollectionId;
+    private String name;
+    private String description;
+    private String guid;
 
-    private String createdBy;
-    private Date createTime;
-
-    private String updatedBy;
-    private Date updateTime;
-
-    private Long version;
-    private Status status;
-
-    private String typeDefName;
-    private String typeDefDescription;
+    private String networkAddress;
+    private String protocol;
+    private String encryptionMethod;
 }
