+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-api/src/main/java/org/odpi/openmetadata/accessservice/assetcatalog/model/Schema.java
@@ -5,31 +5,24 @@ import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-import java.io.Serializable;
-import java.util.List;
 
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.NONE;
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.PUBLIC_ONLY;
 
 /**
- * Relationship object holds properties that are used for displaying details about the classification.
+ * A Schema object provides information about a data asset.
  */
-@EqualsAndHashCode(callSuper = true)
 @JsonAutoDetect(getterVisibility = PUBLIC_ONLY, setterVisibility = PUBLIC_ONLY, fieldVisibility = NONE)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Data
-public class Relationship extends Asset implements Serializable {
+public class Schema {
 
-    private static final long serialVersionUID = 1L;
+    private String name;
+    private String guid;
+    private String encodingStandard;
+    private String versionNr;
+    private String author;
 
-    private String fromEntityPropertyName;
-    private Asset fromEntity;
-
-    private String toEntityPropertyName;
-    private Asset toEntity;
-
-    private List<String> linkedEntities;
+    private String deployedDbSchemaName;
 }
