+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-api/src/main/java/org/odpi/openmetadata/accessservice/assetcatalog/model/Table.java
@@ -1,29 +1,29 @@
 /* SPDX-License-Identifier: Apache-2.0 */
-package org.odpi.openmetadata.accessservice.assetcatalog.model.responses;
+package org.odpi.openmetadata.accessservice.assetcatalog.model;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
-import org.odpi.openmetadata.accessservice.assetcatalog.model.Relationship;
-
-import java.io.Serializable;
-import java.util.List;
 
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.NONE;
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.PUBLIC_ONLY;
 
 /**
- * RelationshipsResponse is the response structure used on the Asset Catalog OMAS REST API calls that
- * returns a relationships object as a response.
+ * Table object contains the properties for a RelationshipTable asset can be find in a database.
  */
-@EqualsAndHashCode(callSuper = true)
 @JsonAutoDetect(getterVisibility = PUBLIC_ONLY, setterVisibility = PUBLIC_ONLY, fieldVisibility = NONE)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Data
-public class RelationshipsResponse extends AssetCatalogOMASAPIResponse implements Serializable {
+public class Table {
 
-    private List<Relationship> relationships;
+    private String name;
+    private String guid;
+    private String typeName;
+    private String owner;
+
+    private String typeVersion;
+    private String typeEncodingStandard;
+    private String typeUsage;
 }
