+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-api/src/main/java/org/odpi/openmetadata/accessservice/assetcatalog/responses/ClassificationsResponse.java
@@ -1,27 +1,29 @@
 /* SPDX-License-Identifier: Apache-2.0 */
-package org.odpi.openmetadata.accessservice.assetcatalog.model;
+package org.odpi.openmetadata.accessservice.assetcatalog.responses;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
+import org.odpi.openmetadata.accessservice.assetcatalog.model.Classification;
+
+import java.io.Serializable;
+import java.util.List;
 
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.NONE;
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.PUBLIC_ONLY;
 
 /**
- * A Schema object provides information about a data asset.
+ * ClassificationsResponse is the response structure used on the Asset Catalog OMAS REST API calls that
+ * returns a classification object as a response.
  */
+@EqualsAndHashCode(callSuper = true)
 @JsonAutoDetect(getterVisibility = PUBLIC_ONLY, setterVisibility = PUBLIC_ONLY, fieldVisibility = NONE)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Data
-public class Schema {
+public class ClassificationsResponse extends AssetCatalogOMASAPIResponse implements Serializable {
 
-    private String name;
-    private String encodingStandard;
-    private String versionNr;
-    private String author;
-
-    private String deployedDbSchemaName;
+    private List<Classification> classifications;
 }
