+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-api/src/main/java/org/odpi/openmetadata/accessservice/assetcatalog/responses/RelationshipsResponse.java
@@ -1,12 +1,12 @@
 /* SPDX-License-Identifier: Apache-2.0 */
-package org.odpi.openmetadata.accessservice.assetcatalog.model.responses;
+package org.odpi.openmetadata.accessservice.assetcatalog.responses;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import org.odpi.openmetadata.accessservice.assetcatalog.model.AssetDescription;
+import org.odpi.openmetadata.accessservice.assetcatalog.model.Relationship;
 
 import java.io.Serializable;
 import java.util.List;
@@ -15,15 +15,15 @@ import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.NONE;
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.PUBLIC_ONLY;
 
 /**
- * AssetDescriptionResponse is the response structure used on the Asset Catalog OMAS REST API calls that returns a
- * asset description object as a response.
+ * RelationshipsResponse is the response structure used on the Asset Catalog OMAS REST API calls that
+ * returns a relationships object as a response.
  */
 @EqualsAndHashCode(callSuper = true)
 @JsonAutoDetect(getterVisibility = PUBLIC_ONLY, setterVisibility = PUBLIC_ONLY, fieldVisibility = NONE)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Data
-public class AssetDescriptionResponse extends AssetCatalogOMASAPIResponse implements Serializable {
+public class RelationshipsResponse extends AssetCatalogOMASAPIResponse implements Serializable {
 
-    private List<AssetDescription> assetDescriptionList;
+    private List<Relationship> relationships;
 }
