+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-client/README.md
@@ -2,6 +2,7 @@
 
 # Asset Catalog Open Metadata Access Service (OMAS) client
 
-The Asset Catalog OMAS supports a REST API for requests. This client
-package provides language-specific client packages to make it easier
+The Asset Catalog OMAS supports a REST API for requests. 
+
+This client package provides language-specific client packages to make it easier
 for data tools and applications to call the interface.
