+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-client/pom.xml
@@ -34,8 +34,14 @@
 
         <dependency>
             <groupId>org.odpi.egeria</groupId>
-            <artifactId>asset-catalog-server</artifactId>
+            <artifactId>asset-catalog-api</artifactId>
             <version>${open-metadata.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-web</artifactId>
+            <version>${spring-web.version}</version>
         </dependency>
 
         <dependency>
