+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-server/README.md
@@ -2,6 +2,10 @@
 
 # Asset Catalog Open Metadata Access Service (OMAS) server-side implementation
 
-The Asset Catalog OMAS server-side support is organized as follows ...
+The Asset Catalog OMAS server-side support is organized as follows:
+* admin package that it is used to initialize and terminate the Asset Catalog OMAS. 
+The initialization call provides this OMAS with resources from the Open Metadata Repository Services.
 
-_... add design info here ..._
+* audit log used to define the message content for the OMRS Audit Log.
+* service package that provides the server-side implementation of the Asset Catalog Open Metadata Assess Service (OMAS).
+These services provide the functionality to fetch asset's header, classification, properties and relationships
