+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-server/src/main/java/org/odpi/openmetadata/accessservice/assetcatalog/util/ExceptionUtil.java
@@ -1,7 +1,7 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 package org.odpi.openmetadata.accessservice.assetcatalog.util;
 
-import org.odpi.openmetadata.accessservice.assetcatalog.model.responses.AssetCatalogOMASAPIResponse;
+import org.odpi.openmetadata.accessservice.assetcatalog.responses.AssetCatalogOMASAPIResponse;
 import org.odpi.openmetadata.repositoryservices.ffdc.exception.OMRSCheckedExceptionBase;
 
 public class ExceptionUtil {
