+++ open-metadata-implementation/access-services/asset-catalog/asset-catalog-spring/src/main/java/org/odpi/openmetadata/accessservice/assetcatalog/server/spring/AssetCatalogEntityResource.java
@@ -4,9 +4,9 @@ package org.odpi.openmetadata.accessservice.assetcatalog.server.spring;
 import org.odpi.openmetadata.accessservice.assetcatalog.model.AssetDescription;
 import org.odpi.openmetadata.accessservice.assetcatalog.model.SequenceOrderType;
 import org.odpi.openmetadata.accessservice.assetcatalog.model.Status;
-import org.odpi.openmetadata.accessservice.assetcatalog.model.responses.AssetDescriptionResponse;
-import org.odpi.openmetadata.accessservice.assetcatalog.model.responses.ClassificationsResponse;
-import org.odpi.openmetadata.accessservice.assetcatalog.model.responses.RelationshipsResponse;
+import org.odpi.openmetadata.accessservice.assetcatalog.responses.AssetDescriptionResponse;
+import org.odpi.openmetadata.accessservice.assetcatalog.responses.ClassificationsResponse;
+import org.odpi.openmetadata.accessservice.assetcatalog.responses.RelationshipsResponse;
 import org.odpi.openmetadata.accessservice.assetcatalog.service.AssetCatalogAssetService;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PathVariable;
